Update ResultCandidatePicker empty state - Rename ApplyButton to InsertButton
