--- packages/scrabble-solver/src/components/Solver/components/ApplyButton/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ApplyButton';
