+++ packages/scrabble-solver/src/components/Solver/components/InsertButton/index.ts
@@ -1,0 +1,1 @@
+export { default } from './InsertButton';
