+++ packages/scrabble-solver/src/components/Solver/components/ResultCandidatePicker/ResultCandidatePicker.module.scss
@@ -90,6 +90,7 @@
   flex: 1;
   text-transform: uppercase;
   text-align: center;
+  white-space: pre-wrap;
 }
 
 .iconContainer {
@@ -113,6 +114,6 @@
   color: var(--color--inactive);
 }
 
-.apply {
+.insert {
   flex: 0 0 auto;
 }
